# Add new `Performance/Sum` cop