--- a/docs/modules/ROOT/pages/cops.adoc
+++ b/docs/modules/ROOT/pages/cops.adoc
@@ -36,6 +36,7 @@
 * xref:cops_performance.adoc#performancestartwith[Performance/StartWith]
 * xref:cops_performance.adoc#performancestringinclude[Performance/StringInclude]
 * xref:cops_performance.adoc#performancestringreplacement[Performance/StringReplacement]
+* xref:cops_performance.adoc#performancesum[Performance/Sum]
 * xref:cops_performance.adoc#performancetimesmap[Performance/TimesMap]
 * xref:cops_performance.adoc#performanceunfreezestring[Performance/UnfreezeString]
 * xref:cops_performance.adoc#performanceuridefaultparser[Performance/UriDefaultParser]