@bbatsov 2020-06-12
Btw, it can also be argued that's a matter of style, not just performance.