@fatkodima 2020-06-12
Updated with:
1. better cop description, as suggested
2. link reference. There is just one non finished PR mentioning `Enumerable#sum` for the whole `fast-ruby`, so I added mentioned link (didn't find better on the internets).